+++ src/main/java/entity/Review.java
@@ -1,11 +1,13 @@
 package entity;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class Review {
     public ReviewId id;
-    public double rating;
+    public BigDecimal rating;
     public String details;
     public String authorId;
 }
